mysite/requirements.txt
africastalking==2.0
amqp==5.3.1
asgiref==3.7.2
autopep8==2.3.2
beautifulsoup4==4.13.5
billiard==4.2.1
Brotli==1.1.0
celery==5.3.6
certifi==2025.8.3
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==46.0.1
cssselect2==0.8.0
Django==4.2.11
django-encrypted-model-fields==0.6.5
django-environ==0.11.2
django-redis==6.0.0
djangorestframework==3.16.1
django-sendgrid-v5==1.3.0
et_xmlfile==2.0.0
flower==2.0.1
fonttools==4.60.0
html5lib==1.1
humanize==4.13.0
idna==3.10
kombu==5.5.4
Levenshtein==0.27.1
lxml==6.0.1
numpy==2.3.3
ofxparse==0.21
openpyxl==3.1.2
packaging==25.0
pandas==2.2.2
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.0
pillow==10.2.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
psycopg2-binary==2.9.9
pycodestyle==2.14.0
pycparser==2.23
pydyf==0.11.0
pypdfium2==4.30.0
pyphen==0.17.2
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-Levenshtein==0.27.1
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.14.1
redis==5.0.3
reportlab==4.1.0
requests==2.32.5
responses==0.25.8
schema==0.7.7
six==1.17.0
soupsieve==2.8
sqlparse==0.4.4
thefuzz==0.22.1
tinycss2==1.4.0
tornado==6.5.2
typing_extensions==4.9.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
weasyprint==62.1
webencodings==0.5.1
xlwt==1.3.0
zopfli==0.2.3.post1
